Add -Wconversion to chip-tool

+++ examples/chip-tool/commands/common/NetworkCommand.h
@@ -61,8 +61,8 @@ private:
 
     char mName[46];
     Command::AddressWithInterface mRemoteAddr;
-    uint32_t mRemotePort;
-    uint32_t mDiscriminator;
+    uint16_t mRemotePort;
+    uint16_t mDiscriminator;
     uint32_t mSetupPINCode;
 };
